moved constants into separate file + added copyleft notices
Having the constants in BatteryMonitor was a bad idea, since when
setup.py is running outside a X environment or the DESKTOP environment
variable is not set (which is what yum forces), then an import gtk
craches the build process. So the new Constants.py does not import
anything to prevent any errors.

Added the gpl 2 text as a constant, and inserted it into the
about-dialog.
Additionally every source file got a copyleft notice on its top.